# duitkita-api/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Duitkita API
  description: >
    Hand-written OpenAPI spec for the Duitkita Go backend, maintained
    separately from the handler code for manual/exploratory testing via
    Swagger UI.


    To try protected endpoints: call `POST /auth/register` or
    `POST /auth/login` below, copy the `access_token` from the response,
    then click **Authorize** and paste it (without the `Bearer ` prefix).
  version: "1.0.0"
servers:
  - url: /api/v1
...
security:
  - bearerAuth: []
paths:
  /auth/register:
    post:
      tags: [Auth]
      summary: Register a new user
      description: >
        Creates an unverified account and emails a 6-digit OTP to the given
        address. No tokens are issued yet — call `POST /auth/verify-otp`
        with the code to complete registration and receive tokens.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RegisterRequest' }
      responses:
        '201':
          description: registered successfully, otp sent
          content:
            application/json:
              schema: { $ref: '#/components/schemas/RegisterResponseEnvelope' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '409': { $ref: '#/components/responses/Conflict' }
  /auth/verify-otp:
    post:
      tags: [Auth]
      summary: Verify the registration OTP and receive tokens
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/VerifyOTPRequest' }
      responses:
        '200':
          description: account verified
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AuthResponseEnvelope' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
  /auth/resend-otp:
    post:
      tags: [Auth]
      summary: Resend an OTP for registration or password reset
      description: >
        Always returns 200 even if the email isn't registered, so the
        response can't be used to enumerate accounts. Subject to a
        per-email cooldown (`OTP_RESEND_COOLDOWN_SECONDS`).
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ResendOTPRequest' }
      responses:
        '200':
          description: otp resent if the account exists
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Envelope' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '409': { $ref: '#/components/responses/Conflict' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
  /auth/forgot-password:
    post:
      tags: [Auth]
      summary: Request a password reset OTP
      description: Always returns 200 even if the email isn't registered.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ForgotPasswordRequest' }
      responses:
        '200':
          description: if the email is registered, a reset code has been sent
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Envelope' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
  /auth/reset-password:
    post:
      tags: [Auth]
      summary: Reset the password using the OTP from /auth/forgot-password
      description: On success, all existing sessions for the user are revoked.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ResetPasswordRequest' }
      responses:
        '200':
          description: password reset successfully
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Envelope' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '404': { $ref: '#/components/responses/NotFound' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
  /auth/login:
    post:
      tags: [Auth]
      summary: Log in with email and password
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/LoginRequest' }
      responses:
        '200':
          description: logged in successfully
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AuthResponseEnvelope' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403':
          description: email not verified yet
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorEnvelope' }
  /auth/refresh:
    post:
      tags: [Auth]
      summary: Exchange a refresh token for a new access/refresh token pair
      description: The refresh token is single-use; each call rotates it.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RefreshTokenRequest' }
      responses:
        '200':
          description: token refreshed
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AuthResponseEnvelope' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /auth/sessions:
    get:
      tags: [Auth]
      summary: List active sessions for the current user
      responses:
        '200':
          description: sessions retrieved
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - type: object
                    properties:
                      data:
                        type: array
                        items: { $ref: '#/components/schemas/SessionResponse' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /auth/sessions/others:
    delete:
      tags: [Auth]
      summary: Revoke every session except the current one
      parameters:
        - in: query
          name: current_session_id
          required: true
          schema: { type: string }
      responses:
        '200': { $ref: '#/components/responses/OK' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /auth/sessions/{id}:
    delete:
      tags: [Auth]
      summary: Revoke one session by id
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200': { $ref: '#/components/responses/OK' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    IdPath:
      in: path
      name: id
      required: true
      schema: { type: string, format: uuid }
...
  responses:
    OK:
      description: success
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Envelope' }
    BadRequest:
      description: validation error
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
    Unauthorized:
      description: missing, malformed or expired token
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
    NotFound:
      description: resource not found, or not owned by the current user
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
    Conflict:
      description: conflicting state (duplicate, already finalized, expired, etc.)
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
    TooManyRequests:
      description: rate limit exceeded, try again later
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
  schemas:
    Envelope:
      type: object
      properties:
        success: { type: boolean, example: true }
        message: { type: string }
        data: {}
      required: [success]
    ErrorEnvelope:
      type: object
      properties:
        success: { type: boolean, example: false }
        error: { type: string }
      required: [success, error]
    AuthResponseEnvelope:
      allOf:
        - $ref: '#/components/schemas/Envelope'
        - type: object
          properties:
            data: { $ref: '#/components/schemas/AuthResponse' }
    RegisterResponseEnvelope:
      allOf:
        - $ref: '#/components/schemas/Envelope'
        - type: object
          properties:
            data: { $ref: '#/components/schemas/RegisterResponse' }
...
    RegisterRequest:
      type: object
      properties:
        name: { type: string, minLength: 2, maxLength: 100, example: Alice }
        email: { type: string, format: email, example: alice@example.com }
        password: { type: string, format: password, minLength: 8, example: password123 }
      required: [name, email, password]
    LoginRequest:
      type: object
      properties:
        email: { type: string, format: email }
        password: { type: string, format: password }
      required: [email, password]
    RefreshTokenRequest:
      type: object
      properties:
        refresh_token: { type: string, description: "Format: <session_id>.<secret>" }
      required: [refresh_token]
    VerifyOTPRequest:
      type: object
      properties:
        email: { type: string, format: email, example: alice@example.com }
        otp: { type: string, minLength: 6, maxLength: 6, example: "042917" }
      required: [email, otp]
    ResendOTPRequest:
      type: object
      properties:
        email: { type: string, format: email }
        purpose: { type: string, enum: [register, reset_password] }
      required: [email, purpose]
    ForgotPasswordRequest:
      type: object
      properties:
        email: { type: string, format: email }
      required: [email]
    ResetPasswordRequest:
      type: object
      properties:
        email: { type: string, format: email }
        otp: { type: string, minLength: 6, maxLength: 6, example: "042917" }
        new_password: { type: string, format: password, minLength: 8 }
      required: [email, otp, new_password]
...
    UserResponse:
      type: object
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        email: { type: string, format: email }
        avatar_url: { type: string, format: uri }
        created_at: { type: string, format: date-time }
    AuthResponse:
      type: object
      properties:
        access_token: { type: string }
        refresh_token: { type: string }
        user: { $ref: '#/components/schemas/UserResponse' }
    RegisterResponse:
      type: object
      properties:
        email: { type: string, format: email }
    SessionResponse:
      type: object
      properties:
        id: { type: string, format: uuid }
        device_name: { type: string }
        ip_address: { type: string }
        user_agent: { type: string }
        last_active_at: { type: string, format: date-time }
        created_at: { type: string, format: date-time }
        is_current: { type: boolean }
...